requirements.txt
discord.py>=2.3.2
sarvamai>=0.1.2
aiohttp>=3.9.5
python-dotenv>=1.0.1
